+++ src/routes/graphql/MutationResolver.ts
@@ -2,7 +2,7 @@ import {GraphQLError} from "graphql";
 import {FileUpload} from "graphql-upload";
 import * as yaml from "js-yaml";
 import isEmail from "validator/lib/isEmail";
-import dataaccess from "../../lib/dataAccess";
+import dataAccess from "../../lib/dataAccess";
 import {BlacklistedError} from "../../lib/errors/BlacklistedError";
 import {GroupNotFoundError} from "../../lib/errors/GroupNotFoundError";
 import {InvalidEmailError} from "../../lib/errors/InvalidEmailError";
@@ -53,7 +53,7 @@ export class MutationResolver extends BaseResolver {
      * @param request
      */
     public async login({email, passwordHash}: { email: string, passwordHash: string }, request: any): Promise<User> {
-        const user = await dataaccess.getUserByLogin(email, passwordHash);
+        const user = await dataAccess.getUserByLogin(email, passwordHash);
         request.session.userId = user.id;
         return user;
     }
@@ -96,7 +96,7 @@ export class MutationResolver extends BaseResolver {
         if (!mailValid) {
             throw new InvalidEmailError(email);
         }
-        const user = await dataaccess.registerUser(username, email, passwordHash);
+        const user = await dataAccess.registerUser(username, email, passwordHash);
         request.session.userId = user.id;
         return user;
     }
@@ -124,8 +124,8 @@ export class MutationResolver extends BaseResolver {
      * @param type
      * @param request
      */
-    public async vote({postId, type}: { postId: number, type: dataaccess.VoteType }, request: any):
-        Promise<{ post: Post, voteType: dataaccess.VoteType }> {
+    public async vote({postId, type}: { postId: number, type: dataAccess.VoteType }, request: any):
+        Promise<{ post: Post, voteType: dataAccess.VoteType }> {
         this.ensureLoggedIn(request);
         const post = await Post.findByPk(postId);
         if (post) {
@@ -143,15 +143,17 @@ export class MutationResolver extends BaseResolver {
      * Creates a new post
      * @param content
      * @param activityId
+     * @param type
      * @param request
      */
-    public async createPost({content, activityId}: { content: string, activityId?: number},
-                            request: any): Promise<Post> {
+    public async createPost(
+        {content, activityId, type}: { content: string, activityId?: number, type: dataAccess.PostType},
+        request: any): Promise<Post> {
         this.ensureLoggedIn(request);
         if (content.length > 2048) {
             throw new GraphQLError("Content too long.");
         }
-        const post = await dataaccess.createPost(content, request.session.userId, activityId);
+        const post = await dataAccess.createPost(content, request.session.userId, activityId, type);
         globals.internalEmitter.emit(InternalEvents.GQLPOSTCREATE, post);
         return post;
     }
@@ -171,7 +173,7 @@ export class MutationResolver extends BaseResolver {
         });
         const isAdmin = (await User.findOne({where: {id: request.session.userId}})).isAdmin;
         if (post.rAuthor.id === request.session.userId || isAdmin) {
-            return await dataaccess.deletePost(post.id);
+            return await dataAccess.deletePost(post.id);
         } else {
             throw new GraphQLError("User is not author of the post.");
         }
@@ -188,7 +190,7 @@ export class MutationResolver extends BaseResolver {
         if (members) {
             chatMembers.push(...members);
         }
-        return await dataaccess.createChat(...chatMembers);
+        return await dataAccess.createChat(...chatMembers);
     }
 
     /**
@@ -200,7 +202,7 @@ export class MutationResolver extends BaseResolver {
     public async sendMessage({chatId, content}: { chatId: number, content: string }, request: any):
         Promise<ChatMessage> {
         this.ensureLoggedIn(request);
-        const message = await dataaccess.sendChatMessage(request.session.userId, chatId, content);
+        const message = await dataAccess.sendChatMessage(request.session.userId, chatId, content);
         globals.internalEmitter.emit(InternalEvents.GQLCHATMESSAGE, message);
         return message;
     }
@@ -211,10 +213,10 @@ export class MutationResolver extends BaseResolver {
      * @param type
      * @param request
      */
-    public async sendRequest({receiver, type}: { receiver: number, type: dataaccess.RequestType }, request: any):
+    public async sendRequest({receiver, type}: { receiver: number, type: dataAccess.RequestType }, request: any):
         Promise<Request> {
         this.ensureLoggedIn(request);
-        return dataaccess.createRequest(request.session.userId, receiver, type);
+        return dataAccess.createRequest(request.session.userId, receiver, type);
     }
 
     /**
@@ -223,7 +225,7 @@ export class MutationResolver extends BaseResolver {
      * @param type
      * @param request
      */
-    public async denyRequest({sender, type}: { sender: number, type: dataaccess.RequestType }, request: any) {
+    public async denyRequest({sender, type}: { sender: number, type: dataAccess.RequestType }, request: any) {
         this.ensureLoggedIn(request);
         const user = await User.findByPk(request.session.userId);
         await user.acceptRequest(sender, type);
@@ -236,7 +238,7 @@ export class MutationResolver extends BaseResolver {
      * @param type
      * @param request
      */
-    public async acceptRequest({sender, type}: { sender: number, type: dataaccess.RequestType }, request: any) {
+    public async acceptRequest({sender, type}: { sender: number, type: dataAccess.RequestType }, request: any) {
         this.ensureLoggedIn(request);
         const user = await User.findByPk(request.session.userId);
         await user.acceptRequest(sender, type);
@@ -262,7 +264,7 @@ export class MutationResolver extends BaseResolver {
      */
     public async createGroup({name, members}: { name: string, members: number[] }, request: any): Promise<Group> {
         this.ensureLoggedIn(request);
-        return await dataaccess.createGroup(name, request.session.userId, members);
+        return await dataAccess.createGroup(name, request.session.userId, members);
     }
 
     /**
@@ -291,8 +293,8 @@ export class MutationResolver extends BaseResolver {
      */
     public async joinGroup({groupId}: { groupId: number }, request: any): Promise<Group> {
         this.ensureLoggedIn(request);
-        return dataaccess.changeGroupMembership(groupId, request.session.userId,
-            dataaccess.MembershipChangeAction.ADD);
+        return dataAccess.changeGroupMembership(groupId, request.session.userId,
+            dataAccess.MembershipChangeAction.ADD);
     }
 
     /**
@@ -302,8 +304,8 @@ export class MutationResolver extends BaseResolver {
      */
     public async leaveGroup({groupId}: { groupId: number }, request: any): Promise<Group> {
         this.ensureLoggedIn(request);
-        return dataaccess.changeGroupMembership(groupId, request.session.userId,
-            dataaccess.MembershipChangeAction.REMOVE);
+        return dataAccess.changeGroupMembership(groupId, request.session.userId,
+            dataAccess.MembershipChangeAction.REMOVE);
     }
 
     /**
@@ -319,8 +321,8 @@ export class MutationResolver extends BaseResolver {
         if (group && !(await group.$has("rAdmins", user)) && (await group.creator()) !== user.id) {
             throw new NotAGroupAdminError(groupId);
         }
-        return dataaccess.changeGroupMembership(groupId, userId,
-            dataaccess.MembershipChangeAction.OP);
+        return dataAccess.changeGroupMembership(groupId, userId,
+            dataAccess.MembershipChangeAction.OP);
     }
 
     /**
@@ -341,8 +343,8 @@ export class MutationResolver extends BaseResolver {
             throw new GraphQLError(
                 "You are not allowed to remove a creator as an admin.");
         }
-        return await dataaccess.changeGroupMembership(groupId, userId,
-            dataaccess.MembershipChangeAction.DEOP);
+        return await dataAccess.changeGroupMembership(groupId, userId,
+            dataAccess.MembershipChangeAction.DEOP);
     }
 
     /**
@@ -361,7 +363,7 @@ export class MutationResolver extends BaseResolver {
         if (!(await group.$has("rAdmins", user))) {
             throw new NotAGroupAdminError(groupId);
         }
-        const blacklisted = await dataaccess.checkBlacklisted(name);
+        const blacklisted = await dataAccess.checkBlacklisted(name);
         if (blacklisted.length > 0) {
             throw new BlacklistedError(blacklisted.map((p) => p.phrase), "event name");
         }